refactor: reorganize and enhance utility tests for MCP resources
